chore: migrate habit types to shared location for improved consistency
